Add tooltips for torque time-series
