+++ src/widgets/constants.js
@@ -1,4 +1,3 @@
 module.exports = {
-  CHART_HEIGHT: 48 + 20 + 4,
   TOOLTIP_TRIANGLE_HEIGHT: 4
 };
