+++ src/widgets/time-series/tooltip.tpl
@@ -1,0 +1,1 @@
+<div class="CDB-Widget-tooltip CDB-Widget-tooltip--light CDB-Text CDB-Size-small js-tooltip"></div>
